sel4

Struct AbsoluteCPtr

pub struct AbsoluteCPtr<C = NoExplicitInvocationContext> { /* private fields */ }

A CPtrWithDepth to be resolved in the context of a particular CNode.

AbsoluteCPtr addresses capability slots in a more general way than Cap. It allows one to address any capability slot that is directly addressable from any CNode that is directly addressible in the current threadโs CSpace. Furthermore, it allows one to address capability slots that contain CNodes by limiting the lookup depth to prevent the kernelโs lookup procedure from descending into the CNode contained in that slot.

seL4_CNode_* capability invocations are methods of AbsoluteCPtr rather than Cap.

In addition to AbsoluteCPtr::new, the following methods can be used to construct an AbsoluteCPtr:

CNode::absolute_cptr

CNode::absolute_cptr_from_bits_with_depth

CNode::absolute_cptr_for_self

Implementations

impl<C> AbsoluteCPtr<C>

pub const fn new(root: CNode<C>, path: CPtrWithDepth) -> Self

pub const fn root(&self) -> &CNode<C>

pub fn into_root(self) -> CNode<C>

pub const fn path(&self) -> &CPtrWithDepth

pub fn with<C1>(self, context: C1) -> AbsoluteCPtr<C1>

pub fn without_context(self) -> AbsoluteCPtr

impl<C: InvocationContext> AbsoluteCPtr<C>

pub fn revoke(self) -> Result<()>

Corresponds to seL4_CNode_Revoke.

pub fn delete(self) -> Result<()>

Corresponds to seL4_CNode_Delete.

pub fn copy(self, src: &AbsoluteCPtr, rights: CapRights) -> Result<()>

Corresponds to seL4_CNode_Copy.

pub fn mint( self, src: &AbsoluteCPtr, rights: CapRights, badge: Word, ) -> Result<()>

Corresponds to seL4_CNode_Mint.

pub fn move_(self, src: &AbsoluteCPtr) -> Result<()>

Corresponds to seL4_CNode_Move.

pub fn mutate(self, src: &AbsoluteCPtr, badge: Word) -> Result<()>

Corresponds to seL4_CNode_Mutate.

pub fn save_caller(self) -> Result<()>

Corresponds to seL4_CNode_SaveCaller.