sel4

Struct Cap

pub struct Cap<T: CapType, C = NoExplicitInvocationContext> { /* private fields */ }

A capability pointer to be resolved in the current CSpace.

The T parameter is a CapType marking the type of the pointed-to capability.

The C parameter is a strategy for discovering the current thread’s IPC buffer. When the "state" feature is enabled, NoExplicitInvocationContext is an alias for ImplicitInvocationContext, which uses the IpcBuffer set by set_ipc_buffer. Otherwise, it is an alias for NoInvocationContext, which does not implement InvocationContext. In such cases, the with method is used to specify an invocation context before the capability is invoked.

The most general way to construct a Cap is with CPtr::cast.

Note that seL4_CNode_* capability invocations are methods of AbsoluteCPtr rather than Cap.

Implementations

impl<C: InvocationContext> Cap<VCpu, C>

pub fn vcpu_set_tcb(self, tcb: Tcb) -> Result<()>

Corresponds to seL4_ARM_VCPU_SetTCB.

pub fn vcpu_read_regs(self, field: VCpuReg) -> Result<Word>

Corresponds to seL4_ARM_VCPU_ReadRegs.

pub fn vcpu_write_regs(self, field: VCpuReg, value: Word) -> Result<()>

Corresponds to seL4_ARM_VCPU_WriteRegs.

pub fn vcpu_ack_vppi(self, irq: Word) -> Result<()>

Corresponds to seL4_ARM_VCPU_AckVPPI.

pub fn vcpu_inject_irq( self, virq: u16, priority: u8, group: u8, index: u8, ) -> Result<()>

Corresponds to seL4_ARM_VCPU_InjectIRQ.

impl<C: InvocationContext> Cap<Smc, C>

pub fn smc_call(self, args: &seL4_ARM_SMCContext) -> Result<seL4_ARM_SMCContext>

impl<T: CapTypeForFrameObject, C: InvocationContext> Cap<T, C>

pub fn frame_map( self, vspace: VSpace, vaddr: usize, rights: CapRights, attrs: VmAttributes, ) -> Result<()>

Corresponds to seL4_ARM_Page_Map.

pub fn frame_unmap(self) -> Result<()>

Corresponds to seL4_ARM_Page_Unmap.

pub fn frame_get_address(self) -> Result<usize>

Corresponds to seL4_ARM_Page_GetAddress.

impl<C: InvocationContext> Cap<PT, C>

pub fn pt_map( self, vspace: VSpace, vaddr: usize, attr: VmAttributes, ) -> Result<()>

impl<C: InvocationContext> Cap<UnspecifiedIntermediateTranslationTable, C>

pub fn generic_intermediate_translation_table_map( self, ty: TranslationTableObjectType, vspace: VSpace, vaddr: usize, attr: VmAttributes, ) -> Result<()>

impl<C: InvocationContext> Cap<IrqControl, C>

pub fn irq_control_get_trigger( self, irq: Word, edge_triggered: bool, dst: &AbsoluteCPtr, ) -> Result<()>

Corresponds to seL4_IRQControl_GetTrigger.

impl<C: InvocationContext> Cap<AsidControl, C>

pub fn asid_control_make_pool( self, untyped: Untyped, dst: &AbsoluteCPtr, ) -> Result<()>

Corresponds to seL4_ARM_ASIDControl_MakePool.

impl<C: InvocationContext> Cap<AsidPool, C>

pub fn asid_pool_assign(self, vspace: VSpace) -> Result<()>

Corresponds to seL4_ARM_ASIDPool_Assign.

impl<T: CapType, C> Cap<T, C>

pub const fn cptr(&self) -> CPtr

pub const fn bits(&self) -> CPtrBits

pub fn cast<T1: CapType>(self) -> Cap<T1, C>

pub fn with<C1>(self, context: C1) -> Cap<T, C1>

pub fn without_context(self) -> Cap<T>

pub fn into_invocation_context(self) -> C

impl<T: CapType> Cap<T>

pub const fn from_cptr(cptr: CPtr) -> Self

pub const fn from_bits(bits: CPtrBits) -> Self

impl<T: CapType, C> Cap<T, C>

pub fn upcast(self) -> Unspecified<C>

impl<C> Cap<Unspecified, C>

pub fn downcast<T: CapType>(self) -> Cap<T, C>

impl<C> Cap<CNode, C>

pub fn absolute_cptr<T: HasCPtrWithDepth>(self, path: T) -> AbsoluteCPtr<C>

Returns the AbsoluteCPtr for path in the context of self.

pub fn absolute_cptr_from_bits_with_depth( self, bits: CPtrBits, depth: usize, ) -> AbsoluteCPtr<C>

Returns the AbsoluteCPtr for CPtrWithDepth::from_bits_with_depth(bits, depth) in the context of self.

pub fn absolute_cptr_for_self(self) -> AbsoluteCPtr<C>

Returns the AbsoluteCPtr for self in its own context.

Currently implemented as:

self.absolute_cptr(CPtrWithDepth::empty())

impl<C: InvocationContext> Cap<Untyped, C>

pub fn untyped_retype( self, blueprint: &ObjectBlueprint, dst: &AbsoluteCPtr, dst_offset: usize, num_objects: usize, ) -> Result<()>

Corresponds to seL4_Untyped_Retype.

impl<C: InvocationContext> Cap<Tcb, C>

pub fn tcb_read_registers( self, suspend: bool, count: Word, ) -> Result<UserContext>

Corresponds to seL4_TCB_ReadRegisters.

pub fn tcb_read_all_registers(self, suspend: bool) -> Result<UserContext>

pub fn tcb_write_registers( self, resume: bool, count: Word, regs: &mut UserContext, ) -> Result<()>

Corresponds to seL4_TCB_WriteRegisters.

pub fn tcb_write_all_registers( self, resume: bool, regs: &mut UserContext, ) -> Result<()>

pub fn tcb_resume(self) -> Result<()>

Corresponds to seL4_TCB_Resume.

pub fn tcb_suspend(self) -> Result<()>

Corresponds to seL4_TCB_Suspend.

pub fn tcb_configure( self, fault_ep: CPtr, cspace_root: CNode, cspace_root_data: CNodeCapData, vspace_root: VSpace, ipc_buffer: Word, ipc_buffer_frame: Granule, ) -> Result<()>

Corresponds to seL4_TCB_Configure.

pub fn tcb_set_space( self, fault_ep: CPtr, cspace_root: CNode, cspace_root_data: CNodeCapData, vspace_root: VSpace, ) -> Result<()>

Corresponds to seL4_TCB_SetSpace.

pub fn tcb_set_sched_params( self, authority: Tcb, mcp: Word, priority: Word, ) -> Result<()>

Corresponds to seL4_TCB_SetSchedParams.

pub fn tcb_set_tls_base(self, tls_base: Word) -> Result<()>

Corresponds to seL4_TCB_SetTLSBase.

pub fn tcb_bind_notification(self, notification: Notification) -> Result<()>

Corresponds to seL4_TCB_BindNotification.

pub fn tcb_unbind_notification(self) -> Result<()>

Corresponds to seL4_TCB_UnbindNotification.

impl<C: InvocationContext> Cap<IrqControl, C>

pub fn irq_control_get(self, irq: Word, dst: &AbsoluteCPtr) -> Result<()>

Corresponds to seL4_IRQControl_Get.

impl<C: InvocationContext> Cap<IrqHandler, C>

pub fn irq_handler_ack(self) -> Result<()>

Corresponds to seL4_IRQHandler_Ack.

pub fn irq_handler_set_notification( self, notification: Notification, ) -> Result<()>

Corresponds to seL4_IRQHandler_SetNotification.

pub fn irq_handler_clear(self) -> Result<()>

Corresponds to seL4_IRQHandler_Clear.

impl<C: InvocationContext> Cap<Endpoint, C>

pub fn send(self, info: MessageInfo)

Corresponds to seL4_Send.

pub fn nb_send(self, info: MessageInfo)

Corresponds to seL4_NBSend.

pub fn recv( self, reply_authority: impl ConveysReplyAuthority, ) -> (MessageInfo, Badge)

Corresponds to seL4_Recv.

pub fn nb_recv( self, reply_authority: impl ConveysReplyAuthority, ) -> (MessageInfo, Badge)

Corresponds to seL4_NBRecv.

pub fn call(self, info: MessageInfo) -> MessageInfo

Corresponds to seL4_Call.

pub fn reply_recv( self, info: MessageInfo, reply_authority: impl ConveysReplyAuthority, ) -> (MessageInfo, Badge)

Corresponds to seL4_ReplyRecv.

pub fn send_with_mrs<T: FastMessages>(self, info: MessageInfo, messages: T)

pub fn recv_with_mrs( self, reply_authority: impl ConveysReplyAuthority, ) -> RecvWithMRs

pub fn call_with_mrs<T: FastMessages>( self, info: MessageInfo, messages: T, ) -> CallWithMRs

impl<C: InvocationContext> Cap<Notification, C>

pub fn signal(self)

Corresponds to seL4_Signal.

pub fn wait(self) -> ((), Badge)

Corresponds to seL4_Wait.

impl<C: InvocationContext> Cap<Tcb, C>

pub fn debug_name(self, name: &[u8])

Corresponds to seL4_DebugNameThread.

impl<T: CapType> Cap<T>

pub fn debug_identify(self) -> u32

Corresponds to seL4_DebugCapIdentify.